refactor(security): Change directory
