--- src/main/java/baedalmate/baedalmate/oauth/annotation/CurrentUser.java
@@ -1,12 +1,0 @@
-package baedalmate.baedalmate.oauth.annotation;
-
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-
-import java.lang.annotation.*;
-
-@Target({ElementType.PARAMETER, ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@AuthenticationPrincipal
-public @interface CurrentUser {
-}
