--- src/main/java/baedalmate/baedalmate/oauth/authentication/OAuth2UserDetails.java
@@ -1,83 +1,0 @@
-package baedalmate.baedalmate.oauth.authentication;
-
-import baedalmate.baedalmate.oauth.SocialType;
-import io.jsonwebtoken.lang.Assert;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
-
-@AllArgsConstructor
-@Builder
-public class OAuth2UserDetails implements UserDetails {
-
-    private SocialType socialType;
-    private String socialId;
-    private String username;
-    @Getter
-    private String email;
-    @Getter
-    private String image;
-    private Set<GrantedAuthority> authorities;
-
-    public SocialType getSocialType() {
-        return socialType;
-    }
-
-    public String getSocialId() {
-        return socialId;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return username;
-    }
-
-    public void setRoles(String... roles) {
-        List<GrantedAuthority> authorities = new ArrayList<>(roles.length);
-
-        for (String role : roles) {
-            authorities.add(new SimpleGrantedAuthority("ROLE_" + role));
-        }
-        this.authorities = Set.copyOf(authorities);
-    }
-
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
