--- src/main/java/baedalmate/baedalmate/oauth/domain/PrincipalDetails.java
@@ -1,90 +1,0 @@
-package baedalmate.baedalmate.oauth.domain;
-
-import baedalmate.baedalmate.domain.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.util.*;
-
-public class PrincipalDetails implements UserDetails, OAuth2User {
-    private static final long serialVersionUID = 1L;
-    private User user;
-    private Map<String, Object> attributes;
-
-    // 일반 시큐리티 로그인시 사용
-    public PrincipalDetails(User user) {
-        this.user = user;
-    }
-
-    // OAuth2.0 로그인시 사용
-    public PrincipalDetails(User user, Map<String, Object> attributes) {
-        this.user = user;
-        this.attributes = attributes;
-    }
-
-    public static PrincipalDetails create(User user) {
-        List<GrantedAuthority> authorities = Collections.
-                singletonList(new SimpleGrantedAuthority("ROLE_USER"));
-        return new PrincipalDetails(user);
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public Long getId() { return user.getId(); }
-
-    @Override
-    public String getPassword() {
-        return "";
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getNickname();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collet = new ArrayList<GrantedAuthority>();
-        collet.add(()->{ return user.getRole();});
-        return collet;
-    }
-    // 리소스 서버로 부터 받는 회원정보
-    @Override
-    public Map<String, Object> getAttributes() {
-        return attributes;
-    }
-
-    // User의 PrimaryKey
-    @Override
-    public String getName() {
-        return user.getId()+"";
-    }
-
-    public String getNickname() {
-        return user.getNickname();
-    }
-}
