--- src/main/java/baedalmate/baedalmate/oauth/dto/TokenResponse.java
@@ -1,11 +1,0 @@
-package baedalmate.baedalmate.oauth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class TokenResponse {
-    private String accessToken;
-    private String refreshToken;
-}
