--- src/main/java/baedalmate/baedalmate/oauth/filter/TokenAuthenticationFilter.java
@@ -1,72 +1,0 @@
-package baedalmate.baedalmate.oauth.filter;
-
-
-import baedalmate.baedalmate.errors.errorcode.UserErrorCode;
-import baedalmate.baedalmate.oauth.provider.TokenProvider;
-import baedalmate.baedalmate.oauth.service.PrincipalDetailsService;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class TokenAuthenticationFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private TokenProvider tokenProvider;
-
-    @Autowired
-    private PrincipalDetailsService principalDetailsService;
-
-    private static final Logger logger = LoggerFactory.getLogger(TokenAuthenticationFilter.class);
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        try {
-            String jwt = getJwtFromRequest(request);
-            if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
-                Long userId = tokenProvider.getUserIdFromToken(jwt);
-                UserDetails userDetails = principalDetailsService.loadUserById(userId);
-                UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-                authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-            }
-        } catch (SecurityException | MalformedJwtException e) {
-            request.setAttribute("exception", UserErrorCode.WRONG_TYPE_TOKEN.getHttpStatus());
-        } catch (ExpiredJwtException e) {
-            request.setAttribute("exception", UserErrorCode.EXPIRED_TOKEN.getHttpStatus());
-        } catch (UnsupportedJwtException e) {
-            request.setAttribute("exception", UserErrorCode.UNSUPPORTED_TOKEN.getHttpStatus());
-        } catch (IllegalArgumentException e) {
-            request.setAttribute("exception", UserErrorCode.WRONG_TOKEN.getHttpStatus());
-        } catch (Exception ex) {
-            logger.error("Could not set user authentication in security context", ex);
-        }
-        filterChain.doFilter(request, response);
-
-    }
-
-    private String getJwtFromRequest(HttpServletRequest request) {
-        String bearerToken = request.getHeader("Authorization");
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7, bearerToken.length());
-        }
-        return null;
-    }
-}
-
