--- src/main/java/baedalmate/baedalmate/oauth/handler/OAuth2LoginFailureHandler.java
@@ -1,23 +1,0 @@
-package baedalmate.baedalmate.oauth.handler;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class OAuth2LoginFailureHandler  extends SimpleUrlAuthenticationFailureHandler {
-
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
-        System.out.println("로그인 실패!!! :" + exception.getMessage());
-        response.sendRedirect("/"); //로그인창으로 이동
-    }
-}
