--- src/main/java/baedalmate/baedalmate/oauth/service/PrincipalDetailsService.java
@@ -1,39 +1,0 @@
-package baedalmate.baedalmate.oauth.service;
-
-import baedalmate.baedalmate.domain.User;
-import baedalmate.baedalmate.oauth.domain.PrincipalDetails;
-import baedalmate.baedalmate.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class PrincipalDetailsService implements UserDetailsService {
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username);
-        System.out.println("username:" + username);
-        if (user == null) {
-            return null;
-        } else {
-            return new PrincipalDetails(user);
-        }
-    }
-
-    @Transactional
-    public UserDetails loadUserById(Long id) {
-        User user = userRepository.findOne(id);
-
-        return PrincipalDetails.create(user);
-    }
-}
