--- src/main/java/baedalmate/baedalmate/oauth/service/strategy/SocialLoadStrategy.java
@@ -1,43 +1,0 @@
-package baedalmate.baedalmate.oauth.service.strategy;
-
-import baedalmate.baedalmate.oauth.dto.OAuth2UserInfo;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.Collections;
-import java.util.Map;
-
-public abstract class SocialLoadStrategy {
-
-    ParameterizedTypeReference<Map<String, Object>> RESPONSE_TYPE  =  new ParameterizedTypeReference<>(){};
-
-    protected final RestTemplate restTemplate = new RestTemplate();
-
-    public OAuth2UserInfo getUserInfo(String accessToken) {
-        HttpHeaders headers = new HttpHeaders();
-
-        setHeaders(accessToken, headers);
-
-        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-
-
-        HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(params, headers);
-
-        return sendRequestToSocialSite(request);//구체 클래스가 구현
-    }
-
-    protected abstract OAuth2UserInfo sendRequestToSocialSite(HttpEntity request);
-
-
-
-    public void setHeaders(String accessToken, HttpHeaders headers) {
-        headers.set("Authorization", "Bearer " + accessToken);
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-        headers.setAccept(Collections.singletonList(MediaType.APPLICATION_JSON));
-    }
-}
