+++ src/main/java/baedalmate/baedalmate/security/annotation/AuthUser.java
@@ -1,0 +1,12 @@
+package baedalmate.baedalmate.security.annotation;
+
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.PARAMETER, ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@AuthenticationPrincipal
+public @interface AuthUser {
+}
