+++ src/main/java/baedalmate/baedalmate/security/logout/CustomLogoutSuccessHandler.java
@@ -1,0 +1,36 @@
+package baedalmate.baedalmate.security.logout;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import org.springframework.http.MediaType;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.http.server.ServletServerHttpResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+public class CustomLogoutSuccessHandler implements LogoutSuccessHandler {
+
+    @Override
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        MappingJackson2HttpMessageConverter jsonConverter = new MappingJackson2HttpMessageConverter();
+        MediaType jsonMimeType = MediaType.APPLICATION_JSON;
+        LogoutResponse logoutResponse = new LogoutResponse(true);
+        if (jsonConverter.canWrite(logoutResponse.getClass(), jsonMimeType)) {
+            jsonConverter.write(logoutResponse, jsonMimeType, new ServletServerHttpResponse(response));
+        }
+    }
+
+    @Data
+    @AllArgsConstructor
+    static class LogoutResponse {
+        private boolean success;
+    }
+}
